haskellPackages.arbtt: 0.10.1 -> 2019-01-27, dontCheck
arbtt was broken for quite a while. Recently fixes for the build failure
were published to master, so I just bumped the source to current master.

Also needed to disable tests to run, as they seem to be broken in an
easy to fix way.

I tested execution of the binaries and they are able to parse my logs.