pius: fix dependencies
`perl` needs to be a native build input to ensure that it is
present in `PATH` during the build, allowing `patchShebangs` to
correctly fix the shebang in `pius-party-worksheet`. Previously,
the shebang was left as `/usr/bin/perl` because `perl` could not
be found by `patchShebangs`.

`six` is a required dependency of `pius` and likely should have
been included in the derivation a long time ago, since the `pius`,
`pius-report` and `pius-keyring-mgr` commands are broken without it.
Annoyingly, `pius` still uses `distutils` rather than `setuptools`
which seems to happily install the package without checking that
its dependencies have been installed.