citrix_receiver: remove old versions
The versions 13.8.0 and 13.9.{0,1} will be EOLed before the end of 19.03
and should be dropped.

To provide an easy upgrade path, all unsupported versions will throw an
evaluation error. All versions that are about the be EOLed can be added
there as well.

For now, all of those deprecated versions are still referenced in
`all-packages.nix`, but should be removed before the next release.

See also https://www.citrix.co.uk/support/product-lifecycle/milestones/receiver.html