pytorch: fix CUDA support
This commit fixes CUDA support when building with `allowUnfree = true` and `cudaSupport = true`. The previous change to pytorch.nix
built, but at runtime cuda support didn't work.

This is a work in progress, the test-suite still doesn't find CUDA, so no CUDA-tests are made of the compiled package.
Also the list of packages in `nativeBuildInputs` and `propagatedBuildInputs` are probably wrong, due to a lack of understanding from my side.