xorg: batch of updates, bump luit
Announcements with details of changes
can be found on the mailing list,
and since IIRC these are all from
this month they all are listed here:

https://lists.x.org/archives/xorg-announce/2019-March/thread.html

-----

Grab luit from mirror because our fetcher's
default behavior doesn't get along with
invisible-island's FTP server apparently (?).

Or perhaps server is problematic, not sure.

Example URL:
ftp://ftp.invisible-island.net/luit/luit-20190106.tgz
* Using curl or wget:
  "Server denied you to change to the given directory"
  (or so)
* Using curl with `--ftp-mode nocwd`: success

curl's manpage suggests this is less-compliant behavior
used for speed, offering a third option that's more
likely to work but not quite as fast.
So it seems like the more compliant behavior is
rejected but the fast behavior is accepted.
Dunno, hopefully it'll be on a content-addressed
server soon enough so it doesn't matter as much :).

Hash used here matches manually-fetched from official URL
as well as from the mirror.