hydra: 2019-02-01 -> 2019-03-18
Yesterday we had a discussion about several Hydra-related problems in the #nixos
IRC channel[1], as a result a lot of stalled PRs were merged.

I think that we want to use the latest Hydra now as it contains lots of
new fixes and folks mentioned that they want to use a self-hosted
Hydra, so they probably use the `nixpkgs` package rather than the
expressions from `NixOS/hydra`.

Simply moving to 2019-03-18 should be fine for now as Hydra doesn't have
a release model.

[1] https://logs.nix.samueldr.com/nixos/2019-03-17#2048416;