Fix cudatoolkit building on non-sandboxed systems
This fixes cudatoolkit building on non-sandboxed system.

The cudatoolkit tests run each of the CUDA binaries as a small smoke
test just to make sure they all can at least somewhat run.

However, the `cuda-gdb` binary doesn't run correctly on non-sandboxed
systems because it picks up versions of Python from /usr/lib.

This PR disables the smoke tests for now.  They should be re-enabled
when we figure out how to make them work for cuda-gdb on non-sandboxed
systems.

This PR is for #57939.