systemd: update debian patches url to snapshots.debian.org
The current approach will fail when enough time has passed. We ideally
want to be reproducible even in a few years of time. So we should pick
the sources of patches wisely as otherwise we can not do that.

(cherry picked from commit bb821c65ff1403d07aaf86a46098ca618c5d1aa5)