password-store: link ext. man pages into the password-store $out
When pass was installed with extensions in the system environment, the
man pages for the selected extensions were not available globally
because they were only available in a buildInput of the password-store
derivation.

This commit resolves the problem by linking the man pages from the
extensions environment into the output directory of the password-store
derivation.

Bug 56850