firefox/wrapper.nix: make firefox user configurable
This commit changes the way the firefox wrapper works to allow passing
system wide configuration to the firefox instance, in the form of an
enterprise policies file (policies.json) and/or an AutoConfig file.

Since firefox essentially hard codes the paths to these configuration
files, and due to the immutable nature of the nix store, we have to
copy/symlink the entire runtime.
Also since firefox resolves those paths relative to the "true" path of
the binary (i.e. following symlinks), the binary itself has to be copied.