nixos/nzbget: cfg.configFile/cfg.dataDir owned by cfg.user/cfg.group
These changes ensure, via the preStart script, that the configFile and
dataDir are owned by the user & group that nzbget has been configured
to run with.

Without this, when the config template is copied over to facilitate
initial startup, it's done so with root permissions. This causes the
service to fail during the startup script (grep is used to inspect the
file for some values). The dataDir is similar circumstances; without
these changes, it's created with root permissions, which means the
service cannot read or write data there, if it's running as a
different user.