nixos/nzbget: cfg.configFile should be owned by cfg.user/cfg.group
These changes ensure, via the preStart script, that the configFile is
owned by the user & group that nzbget has been configured to run with.
Without this, when the config template is copied over to facilitate
initial startup, it's done so with root permissions. This causes the
service to fail during the startup script (grep is used to inspect the
file for some values).