nixos/nzbget: cfg.dataDir should be owned by cfg.user/cfg.group
These changes ensure, via the preStart script, that the dataDir is
present and owned by the user & group that nzbget has been configured
to run with.

Without this, the service fails during the startup script and data is
unable to be written to the directory, if the directory isn't owned by
the user under which the service is running.