pmdk: init at 1.6
Intel is finally shipping real persistent memory, now that Optane DCPMMs
are available with Cascade Lake processors. Therefore, programmers need
a persistent memory programming API!

In particular, pmdk is needed for QEMU features relating to persistent
memory: by enabling pmdk as a dependency of QEMU, you can proxy NVDIMMs
from the host system to virtual machines with the exact same consistency
guarantees. (In the normal case, these host NVDIMMs can be used as a
backend device for DAX-enabled filesystems, and the persistent memory
given to the virtual machine can be represeted as objects in the
filesystem, allowing granular distribution of non-volatile memory to
clients.)

Signed-off-by: Austin Seipp <aseipp@pobox.com>