build-fhs-userenv: make use of dynamic linker config
This ensures the right order. When using LD_LIBRARY_PATH, /usr/lib is
checked *first* instead of last (which is normal behaviour on non-nixos
systems).

This can lead to issues. One example is a self-compiled python, which
links to libpython2.7.so. It may have /path/to/custom/compiled/python in
its `runpath`, but it will link to /usr/lib/libpython2.7 anyway since
/usr/lib is in LD_LIBRARY_PATH. This means the wrong python version will
be executed, leading to very confusing behavior.