Rewrite debian-build
This commit rewrites the debian-build expression because it was
cluttered with uneccessary attribute sets.

All sets were merged into one set.

The attribute "name" was renamed so it can more easily be distinguished
from the "name" attribute of the resulting set.

This somehow fixes the build of debian packages in
`pkgs.buildRecDebpkg.makeDeb`. I don't know why.